another cleanup
hopefully I've merged everything into the develop branch
